cppcheck fixes

+++ src/poller.cpp
@@ -90,7 +90,7 @@ int ZmqPoller::init(Pollable const& p, zmq_pollitem_t& item) noexcept
 
 int ZmqPoller::doPoll(int timeout_ms, typename base::PollItemList& items) noexcept
 {
-	zth_assert(items.size() <= std::numeric_limits<int>::max());
+	zth_assert(items.size() <= (size_t)std::numeric_limits<int>::max());
 	int res = zmq_poll(&items[0], (int)items.size(), timeout_ms);
 	if(res < 0)
 		return zmq_errno();
